[FIRRTL][LowerTypes] Extend SubaccessOp hack
The subaccess op is not supported yet in the lower-types pass. To get
around this limitation, there is an incorrect lowering of this
operation.  The subaccess op is lowered to always return the first index
of the array.

This change extends the hack to use support full lowering when there is
a vector of bundles.  It does this by creating a Subindex[0] op, and
lowering that.

This also change the diagnostic from an `error` to a `warning`.