[FIRRTL] Annotate GCT Companion with "circt.bind"
Remove annotations marking Grand Central companion Modules and generate an
annotation, "circt.bind", on each of their Instances.  This is a crappy way of
indicating instances that need to be lowered to sv::BindOp later, but without
adding a bind op to the FIRRTL Dialect.  There are benefits to doing it this way
that enable Chisel to access this functionality via annotations without needing
to change how it emits CHIRRTL.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>